requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
python-dotenv==1.0.0
firebase-admin==6.3.0
stripe==7.7.0
gunicorn==21.2.0
# psycopg2-binary==2.9.9  # Désactivé pour Python 3.13 - à réactiver pour Railway avec PostgreSQL
Flask-CORS==4.0.0
Werkzeug==3.0.1
requests==2.31.0
